correction int => size_t + 0 warning

+++ HeapSort.c
@@ -9,30 +9,30 @@
 #include <limits.h>
 #include "Sort.h"
 
-static void swap(int *array, int i, int j);
-static int Left(int i);
-static int Right(int i);
-static void Max_Heapify(int *array, int i, size_t heap_size);
+static void swap(int *array, size_t i, size_t j);
+static size_t Left(size_t i);
+static size_t Right(size_t i);
+static void Max_Heapify(int *array, size_t i, size_t heap_size);
 static void Build_Max_Heap(int *array, size_t length);
 
-static int Left(int i){
+static size_t Left(size_t i){
   return 2*i;
 }
 
-static int Right(int i){
+static size_t Right(size_t i){
   return ((2*i)+1);
 }
 
-static void swap(int *array, int i, int j){
+static void swap(int *array, size_t i, size_t j){
   int tmp = array[i];
   array[i] = array[j];
   array[j] = tmp;
 }
 
-static void Max_Heapify(int *array, int i, size_t heap_size){
-  int largest = 0;
-  int l = Left(i);
-  int r = Right(i);
+static void Max_Heapify(int *array, size_t i, size_t heap_size){
+  size_t largest = 0;
+  size_t l = Left(i);
+  size_t r = Right(i);
 
   if(l <= heap_size && array[l] > array[i])
     largest = l;
@@ -52,7 +52,7 @@ static void Max_Heapify(int *array, int i, size_t heap_size){
 static void Build_Max_Heap(int *array, size_t length){
   size_t heap_size = length;
 
-  for(int i = length/2; i >= 0; i--)
+  for(int i = length/2; i >= 0; i--) // int parce que à la fin de la boucle i = -1
     Max_Heapify(array, i, heap_size);
 }
 
@@ -60,7 +60,7 @@ static void HeapSort(int *array, size_t length){
   Build_Max_Heap(array, length);
   size_t heap_size = length;
 
-  for(int i = length; i >= 1; i--){
+  for(size_t i = length; i >= 1; i--){
     swap(array, i , 0);
     heap_size--;
     Max_Heapify(array, 0, heap_size);
